requirements.txt
tensorflow==2.15.0
numpy
pandas
scikit-learn
matplotlib
seaborn
nltk
datasets
transformers
tqdm
keras
flask
pdfplumber
python-docx
